# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: runtime
    description: Colab runtime control operations
  - name: assistant
    description: AI assistant operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /runtime/status:
    get:
      operationId: getRuntimeStatus
      tags: [runtime]
      summary: Get the current Colab runtime status
      responses:
        "200":
          description: Runtime status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeStatus"
  /runtime/bootstrap:
    post:
      operationId: createRuntimeBootstrap
      tags: [runtime]
      summary: Create a temporary Colab/Kaggle connector session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapInput"
      responses:
        "201":
          description: Colab connector instructions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapResponse"
  /runtime/disconnect:
    post:
      operationId: disconnectRuntime
      tags: [runtime]
      summary: Disconnect the active Colab runtime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionInput"
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeStatus"
  /runtime/execute:
    post:
      operationId: executeRuntimeCode
      tags: [runtime]
      summary: Queue Python code for the connected Colab runtime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteInput"
      responses:
        "202":
          description: Code queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResponse"
  /runtime/interrupt:
    post:
      operationId: interruptRuntime
      tags: [runtime]
      summary: Ask the Colab runtime to interrupt execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionInput"
      responses:
        "202":
          description: Interrupt queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResponse"
  /runtime/events:
    get:
      operationId: getRuntimeEvents
      tags: [runtime]
      summary: Read runtime events after a cursor
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Runtime event stream snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeEventsResponse"
  /colab/connect:
    post:
      operationId: connectColabRuntime
      tags: [runtime]
      summary: Register a Colab connector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectInput"
      responses:
        "200":
          description: Connected runtime status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeStatus"
  /colab/events:
    post:
      operationId: postColabEvent
      tags: [runtime]
      summary: Post an event from the Colab connector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ColabEventInput"
      responses:
        "202":
          description: Event accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptedResponse"
  /colab/commands:
    get:
      operationId: getColabCommands
      tags: [runtime]
      summary: Fetch pending commands for the Colab connector
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pending connector commands
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColabCommandsResponse"
  /assistant/chat:
    post:
      operationId: sendAssistantMessage
      tags: [assistant]
      summary: Ask the CC R2 agent to interpret a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantInput"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantResponse"
  /assistant/chat/thread:
    get:
      operationId: getChatThread
      tags: [assistant]
      summary: Get the current chat thread
      responses:
        "200":
          description: Chat thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatThreadResponse"
  /occupancy:
    get:
      operationId: getOccupancy
      tags: [runtime]
      summary: Get single-user occupancy state
      responses:
        "200":
          description: Occupancy state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OccupancyState"
  /memory:
    get:
      operationId: getMemory
      tags: [assistant]
      summary: Get the session memory (messages, plan, decisions, points, tasks)
      responses:
        "200":
          description: Memory summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemorySummary"
  /memory/plan:
    post:
      operationId: addPlanItem
      tags: [assistant]
      summary: Add a plan item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanItemInput"
      responses:
        "201":
          description: Plan item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanItem"
    delete:
      operationId: clearPlan
      tags: [assistant]
      summary: Clear the plan
      responses:
        "200":
          description: Plan cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemorySummary"
  /memory/plan/{id}/toggle:
    post:
      operationId: togglePlanItem
      tags: [assistant]
      summary: Toggle a plan item's done state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated plan item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanItem"
  /memory/points:
    post:
      operationId: addPoint
      tags: [assistant]
      summary: Add an important point
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanItemInput"
      responses:
        "201":
          description: Point added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NamedItem"
  /memory/decisions:
    post:
      operationId: addDecision
      tags: [assistant]
      summary: Record a decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanItemInput"
      responses:
        "201":
          description: Decision added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NamedItem"
  /projects/save:
    post:
      operationId: saveProject
      tags: [assistant]
      summary: Save the current project to GitHub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectSaveInput"
      responses:
        "201":
          description: Project saved to GitHub
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSaveResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    BootstrapInput:
      type: object
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 80
        target:
          type: string
          enum: [colab, kaggle]
      required:
        - label
    SessionInput:
      type: object
      properties:
        sessionId:
          type: string
          minLength: 8
      required:
        - sessionId
    RuntimeStatus:
      type: object
      properties:
        state:
          type: string
          enum: [offline, waiting, connected, busy, error]
        sessionId:
          type: ["string", "null"]
        label:
          type: ["string", "null"]
        target:
          type: ["string", "null"]
          enum: [colab, kaggle, null]
        connectedAt:
          type: ["string", "null"]
        lastSeenAt:
          type: ["string", "null"]
        queuedCommands:
          type: number
        pythonVersion:
          type: ["string", "null"]
      required: [state, sessionId, label, target, connectedAt, lastSeenAt, queuedCommands, pythonVersion]
    BootstrapResponse:
      type: object
      properties:
        sessionId:
          type: string
        token:
          type: string
        connectorCode:
          type: string
        expiresAt:
          type: string
      required: [sessionId, token, connectorCode, expiresAt]
    ConnectInput:
      type: object
      properties:
        sessionId:
          type: string
        token:
          type: string
        runtimeName:
          type: string
          maxLength: 120
        pythonVersion:
          type: string
          maxLength: 80
      required: [sessionId, token]
    ExecuteInput:
      type: object
      properties:
        sessionId:
          type: string
        code:
          type: string
          minLength: 1
          maxLength: 30000
        description:
          type: ["string", "null"]
      required: [sessionId, code]
    ExecutionResponse:
      type: object
      properties:
        accepted:
          type: boolean
        commandId:
          type: string
        message:
          type: string
      required: [accepted, commandId, message]
    RuntimeEvent:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [status, stdout, stderr, result, error, command, system]
        message:
          type: string
        payload:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, type, message, payload, createdAt]
    RuntimeEventsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/RuntimeEvent"
        cursor:
          type: string
      required: [events, cursor]
    ColabEventInput:
      type: object
      properties:
        sessionId:
          type: string
        token:
          type: string
        type:
          type: string
          enum: [status, stdout, stderr, result, error, system]
        message:
          type: string
          maxLength: 30000
        payload:
          type: ["string", "null"]
        commandId:
          type: ["string", "null"]
      required: [sessionId, token, type, message]
    ColabCommand:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, code, description, createdAt]
    ColabCommandsResponse:
      type: object
      properties:
        commands:
          type: array
          items:
            $ref: "#/components/schemas/ColabCommand"
      required: [commands]
    AcceptedResponse:
      type: object
      properties:
        accepted:
          type: boolean
        message:
          type: string
      required: [accepted, message]
    AssistantInput:
      type: object
      properties:
        sessionId:
          type: ["string", "null"]
        message:
          type: string
          minLength: 1
          maxLength: 12000
        preference:
          type: string
          enum: [primary, fast]
        execute:
          type: boolean
      required: [message, execute]
    AssistantResponse:
      type: object
      properties:
        reply:
          type: string
        code:
          type: ["string", "null"]
        commandId:
          type: ["string", "null"]
        provider:
          type: string
        model:
          type: string
      required: [reply, code, commandId, provider, model]
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        code:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, role, content, code, createdAt]
    ChatThreadResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
      required: [messages]
    MemorySummary:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        plan:
          type: array
          items:
            $ref: "#/components/schemas/PlanItem"
        decisions:
          type: array
          items:
            $ref: "#/components/schemas/NamedItem"
        points:
          type: array
          items:
            $ref: "#/components/schemas/NamedItem"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/ProjectTask"
      required: [messages, plan, decisions, points, tasks]
    PlanItem:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        done:
          type: boolean
        createdAt:
          type: string
      required: [id, text, done, createdAt]
    NamedItem:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        createdAt:
          type: string
      required: [id, text, createdAt]
    ProjectTask:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        githubRepo:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, githubRepo, createdAt, updatedAt]
    OccupancyState:
      type: object
      properties:
        busy:
          type: boolean
        ownerId:
          type: ["string", "null"]
      required: [busy, ownerId]
    ProjectSaveInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
      required: [name]
    ProjectSaveResponse:
      type: object
      properties:
        repo:
          type: string
        commitSha:
          type: string
        taskId:
          type: string
      required: [repo, commitSha, taskId]
    PlanItemInput:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
      required: [text]