# lib/api-zod/src/generated/types/bootstrapInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BootstrapInputTarget } from './bootstrapInputTarget';

export interface BootstrapInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  label: string;
  target?: BootstrapInputTarget;
}