# lib/api-zod/src/generated/types/memorySummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';
import type { NamedItem } from './namedItem';
import type { PlanItem } from './planItem';
import type { ProjectTask } from './projectTask';

export interface MemorySummary {
  messages: ChatMessage[];
  plan: PlanItem[];
  decisions: NamedItem[];
  points: NamedItem[];
  tasks: ProjectTask[];
}